Renaming from wrapper_helpers to arithmetic_operation_wrapper
